+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Generate an error if a view contains a `@ScheduleRender` annotated method and is not of type `STATEFUL`.
+
 ### [v0.168](https://github.com/react4j/react4j/tree/v0.168) (2020-06-25) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.167...v0.168)
 
 Changes in this release: